+++ src/main/java/com/akiba/backend/auction/dto/response/AuctionPostSimpleResponse.java
@@ -6,8 +6,11 @@
 package com.akiba.backend.auction.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -15,8 +18,8 @@ public class AuctionPostSimpleResponse {
     private Long postId;
     private String title;
     private String specialType;
-    private int startPrice;            // ← 이 필드 추가
-    private Integer currentPrice;      // currentHighestBid → currentPrice
+    private BigDecimal startPrice;            // ← 이 필드 추가
+    private BigDecimal currentPrice;      // currentHighestBid → currentPrice
     private int bidCount;
     private Integer viewCount;         // 인기 경매에서만 사용
     private String thumbnailUrl;
